+++ src/index.js
@@ -37,5 +37,5 @@ export {
   createOMETiffLoader,
   createZarrLoader,
   createBioformatsZarrLoader,
-  HTTPStore
+  HTTPStore,
 };