+++ src/loaders/decoder.js
@@ -1,4 +1,4 @@
-import { getDecoder } from 'geotiff';
+import { getDecoder } from 'geotiff/src/compression';
 
 export function getTiffDecoder(fileDirectory) {
   return getDecoder(fileDirectory);