+++ tests/loaders/zarrLoader.spec.js
@@ -1,11 +1,11 @@
-/* eslint-disable import/no-extraneous-dependencies, no-unused-expressions */
-import { expect } from 'chai';
+import test from 'tape';
 
 import { zeros, NestedArray } from 'zarr';
 import ZarrLoader from '../../src/loaders/zarrLoader';
 
-describe('Test zarr non-rgb image loader', () => {
-  it('Properties test', async () => {
+test('ZarrLoader Properties test', async t => {
+  t.plan(3);
+  try {
     const z = await zeros([2, 500, 250], {
       chunks: [1, 100, 100],
       dtype: '<f4'
@@ -20,12 +20,18 @@ describe('Test zarr non-rgb image loader', () => {
     ];
     const loader = new ZarrLoader({ data: z, dimensions });
     const { dtype, numLevels, tileSize } = loader;
-    expect(dtype).to.equal('<f4');
-    expect(numLevels).to.equal(1);
-    expect(tileSize).to.equal(100);
-  });
+    t.equal(dtype, '<f4');
+    t.equal(numLevels, 1);
+    t.equal(tileSize, 100);
+    t.end();
+  } catch (e) {
+    t.fail(e);
+  }
+});
 
-  it('getTile tests', async () => {
+test('ZarrLoader getTile tests', async t => {
+  t.plan(3);
+  try {
     const z = await zeros([2, 500, 250], {
       chunks: [1, 100, 100],
       dtype: '<f4'
@@ -45,22 +51,27 @@ describe('Test zarr non-rgb image loader', () => {
       y: 0,
       loaderSelection: [{ channel: 0 }]
     });
-    expect(res.data[0][10]).to.equal(42);
-    expect(res.data[0].length).to.equal(tileSize * tileSize);
+    t.equal(res.data[0][10], 42);
+    t.equal(res.data[0].length, tileSize * tileSize);
 
     res = await loader.getTile({
       x: 0,
       y: 0,
       loaderSelection: [{ channel: 1 }]
     });
-    expect(res.data[0].subarray(0, 100)).to.deep.equal(
+    t.deepEqual(
+      res.data[0].subarray(0, 100),
       NestedArray.arange(100, '<f4').flatten()
     );
-  });
+    t.end();
+  } catch (e) {
+    t.fail(e);
+  }
 });
 
-describe('Test zarr pyramid', () => {
-  it('Pyramid tiles have correct dimensions', async () => {
+test('Pyramid tiles have correct dimensions', async t => {
+  t.plan(2);
+  try {
     const z0 = await zeros([4, 100, 150], {
       chunks: [1, 10, 10],
       dtype: '<f4'
@@ -91,16 +102,18 @@ describe('Test zarr pyramid', () => {
     const { data: level1Tiles } = await loader.getTile({ ...config, z: 1 });
     const { data: level2Tiles } = await loader.getTile({ ...config, z: 2 });
 
-    expect([
-      baseTiles.length,
-      level1Tiles.length,
-      level2Tiles.length
-    ]).to.deep.equal([1, 1, 1]);
+    t.deepEqual(
+      [baseTiles.length, level1Tiles.length, level2Tiles.length],
+      [1, 1, 1]
+    );
 
-    expect([
-      baseTiles[0].length,
-      level1Tiles[0].length,
-      level2Tiles[0].length
-    ]).to.deep.equal([100, 100, 100]);
-  });
+    t.deepEqual(
+      [baseTiles[0].length, level1Tiles[0].length, level2Tiles[0].length],
+      [100, 100, 100]
+    );
+
+    t.end();
+  } catch (e) {
+    t.fail(e);
+  }
 });